requirements.txt
Flask
Gunicorn
requests
python-dotenv
authlib
pymongo
sendgrid